UBUNTU: SAUCE: Add compat macro for skb_get_hash
Commit df4530d690a14dd net: Change skb_get_rxhash to skb_get_hash
changed a function that is part of the 3.13 kernel API to pull in
some newer driver code.
In order for external modules (which check the LINUX_VERSION_CODE
to find out which API functions to use) to still compile correctly,
add a compat definition for the old name.

BugLink: http://bugs.launchpad.net/bugs/1358162

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Andy Whitcroft <andy.whitcroft@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>